code clean up for operand stack versions
